Fully converts ScriptReading to grammar
Replaces the keyword search with a grammar based search. This seems to
speed up the process of finding multiple words.
However, watching the log shows that it is still terribly inaccurate,
'hello' sometimes gets recognized as 'pizza' somehow. Pruned down the
dictionary to only our vocabulary and renamed the full dictionary to
cmudict-en-us-old.dict so we can add more words if necessary.